refactor: move all graphic elements to a dedicated module
add the gui module for graphic elements
add several factory functions for callbacks and graphic objects/widgets creation

close issue #3
